refactor(agent-core): the per-kind matchers live beside the gate, not inside it
The file-size scan refused permission-gate.ts at 447 lines (the 300-line
anti-monolith rule). The gate owns WHEN a pattern is consulted — deny,
unevaluable, allow, mode policy; argument-matchers.ts owns HOW one
pattern is compared with one argument of a declared kind. The one
declared degradation — a path segment that does not percent-decode is
reported as unevaluable, never swallowed — now carries its
allow-fallback reason, which the no-fallback scan asked for.

Lane: L2
